add forgot/reset password flow with email + JWT expiry fixes

+++ backend/src/main/java/com/aurora/info_hub/entity/PasswordResetToken.java
@@ -7,7 +7,6 @@ import lombok.Data;
 import lombok.NoArgsConstructor;
 
 import java.time.LocalDateTime;
-import java.util.List;
 
 @Entity
 @Builder
@@ -18,7 +17,7 @@ import java.util.List;
 public class PasswordResetToken {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private int id;
+    private Long id;
 
     @Column(nullable = false,unique = true)
     private String token;
